Example 모듈 Member 모듈로 교체 필요 : refactor : 포매팅 처리 https://github.com/TEAM-Tripgether/Tripgether-BE/issues/19
