+++ TG-Common/build.gradle
@@ -29,6 +29,11 @@ dependencies {
     
     // Database
     api 'org.postgresql:postgresql:42.7.7'
+
+    // Suh Libraries
+    api "me.suhsaechan:suh-random-engine:1.1.0"
+    api "me.suhsaechan:suh-logger:1.3.11"
+    api "me.suhsaechan:suh-api-log:1.1.5"
     
     // Development tools
     api 'org.springframework.boot:spring-boot-devtools'
