+++ TG-Web/src/main/java/com/tripgether/web/TripgetherApplication.java
@@ -1,4 +1,4 @@
-package com.tripgether;
+package com.tripgether.web;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,9 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class TripgetherApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(TripgetherApplication.class, args);
-	}
-
+    public static void main(String[] args) {
+        SpringApplication.run(TripgetherApplication.class, args);
+    }
 }
-
