+++ TG-Web/src/main/java/com/tripgether/web/config/ComponentScanConfig.java
@@ -3,12 +3,7 @@ package com.tripgether.web.config;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
-/**
- * 컴포넌트 스캔 설정
- * 모든 모듈의 패키지를 스캔하여 Bean으로 등록
- */
+/** 컴포넌트 스캔 설정 모든 모듈의 패키지를 스캔하여 Bean으로 등록 */
 @Configuration
 @ComponentScan(basePackages = "com.tripgether")
-public class ComponentScanConfig {
-}
-
+public class ComponentScanConfig {}
